requirements.txt
openai==0.27.4
datetime
argparse
pandas
flask==2.2
flask_sqlalchemy
gunicorn
PyYAML==6.0
pymysql
pymssql
pyodbc
sqlalchemy
cloud-sql-python-connector==1.2.2
mysql-connector-python
azure-identity
azure-keyvault-secrets
Werkzeug==2.2.2
bcrypt